requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
audioread==3.0.1
bcrypt==4.3.0
billiard==4.2.1
boto3==1.34.0
botocore==1.34.162
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
decorator==5.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
essentia==2.1b6.dev1389
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpickle==4.1.1
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
minio==7.2.0
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
music21==9.7.1
mutagen==1.47.0
networkx==3.4.2
numba==0.61.2
numpy==1.24.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
passlib==1.7.4
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
proglog==0.1.12
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
requests==2.32.5
rsa==4.9.1
s3transfer==0.9.0
scikit-learn==1.7.2
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==1.0.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
websockets==15.0.1
stripe==7.8.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-watch==4.2.0
